backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlmodel==0.0.22
pydantic==2.9.2
python-dotenv==1.0.1
openai==1.51.2
google-generativeai>=0.8.3
asyncpg==0.30.0
alembic==1.13.3
pydantic-settings==2.6.1
bcrypt==3.2.0
passlib[bcrypt]==1.7.4